CCBC-1395: update config revision to use int64_t
Reviewed-on: http://review.couchbase.org/c/libcouchbase/+/152798
Tested-by: Build Bot <build@couchbase.com>

+++ src/http/http-priv.h
@@ -277,7 +277,7 @@ struct Request {
      * Last revision ID of vBucket config. If the current revision does not
      * match this number, the ::used_nodes field is cleared
      */
-    int last_vbcrev;
+    int64_t last_vbcrev;
 
     const lcb_HTTP_TYPE reqtype; /**< HTTP API type */
